Refactor widget expression parsing to its own mixin
